Split commands module
