+++ src/commands/mod.rs
@@ -1,0 +1,7 @@
+mod events;
+mod feeds;
+mod import;
+
+pub use events::{add_event, delete_event, list_events};
+pub use feeds::{generate_ical, generate_rss};
+pub use import::import_ics;
